parameter type matching

+++ src/commons/LocalParameters.h
@@ -150,9 +150,9 @@ public:
     int inputFormat;
     int pdbOutputMode;
 
-    double initEPS = 0.1;
-    double deltaEPS = 0.1;
-    double thresholdCV = 0.1;
+    float initEPS = 0.1;
+    float deltaEPS = 0.1;
+    float thresholdCV = 0.1;
 
     static std::vector<int> getOutputFormat(int formatMode, const std::string &outformat, bool &needSequences, bool &needBacktrace, bool &needFullHeaders,
                                             bool &needLookup, bool &needSource, bool &needTaxonomyMapping, bool &needTaxonomy, bool &needQCa, bool &needTCa, bool &needTMaligner,
